python/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "dlt645"
version = "1.3.2"
description = "DLT645协议Python实现库"
readme = "README.md"
requires-python = ">=3.9"
license = { text = "Apache License 2.0" }
authors = [{ name = "Chen Dongyu", email = "1755696012@qq.com" }]
maintainers = [{ name = "Chen Dongyu", email = "1755696012@qq.com" }]
keywords = ["dlt645", "protocol", "communication", "energy", "meter"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Communications",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
]
dependencies = ["loguru>=0.7.0", "pyserial>=3.5"]

[project.optional-dependencies]
dev = [
    "pytest>=6.0",
    "pytest-cov>=2.0",
    "black>=21.0",
    "flake8>=3.8",
    "mypy>=0.800",
]

[project.urls]
Homepage = "https://gitee.com/chen-dongyu123/dlt645"
Repository = "https://gitee.com/chen-dongyu123/dlt645"
Issues = "https://gitee.com/chen-dongyu123/dlt645/issues"
Changelog = "https://gitee.com/chen-dongyu123/dlt645/tree/master/python/CHANGELOG.md"
Documentation = "https://gitee.com/chen-dongyu123/dlt645/wiki"

[tool.setuptools]
packages = ["dlt645"]
package-dir = { "dlt645" = "src" }
include-package-data = true

[tool.setuptools.package-data]
"dlt645" = ["config/*.json"]

[tool.pytest.ini_options]
testpaths = ["test"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--strict-config",
    "--cov=src",
    "--cov-report=term-missing:skip-covered",
    "--cov-report=html:htmlcov",
    "--cov-report=xml",
]

[tool.black]
line-length = 88
target-version = ["py39", "py310", "py311", "py312", "py313"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = ["serial.*", "loguru.*"]
ignore_missing_imports = true
